+++ src/app/app.component.ts
@@ -1,10 +1,15 @@
-import { Component } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent {
-  title = 'homework-angular';
+export class AppComponent implements AfterViewInit {
+  title = 'Angular Training - Task 2';
+  @ViewChild('appTitle') titleElement: ElementRef<HTMLHeadingElement>;
+
+  ngAfterViewInit() {
+    this.titleElement.nativeElement.innerText = this.title;
+  }
 }