+++ src/app/app.component.ts
@@ -9,7 +9,7 @@ export class AppComponent implements AfterViewInit {
   title = 'Angular Training - Task 2';
   @ViewChild('appTitle') titleElement: ElementRef<HTMLHeadingElement>;
 
-  ngAfterViewInit() {
+  ngAfterViewInit(): void {
     this.titleElement.nativeElement.innerText = this.title;
   }
 }