+++ changelog.md
@@ -22,3 +22,4 @@ Changed Product component; Moved all logic to its parent - ProductList
 30.01.2021
 Added CartItem component; Moved all logic to its parent; Refactored and extended Cart service; Use OnPush strategy wher possible; Refactored models a little
 Added DOM events into CartItem and Product components
+Added header into App component and set its text via ViewChild