+++ changelog.md
@@ -24,3 +24,6 @@ Added CartItem component; Moved all logic to its parent; Refactored and extended
 Added DOM events into CartItem and Product components
 Added header into App component and set its text via ViewChild
 Added highlight directive; TSLint fixes
+
+31.01.2021
+Modified Cart service and modified components which are using it accordingly